Separated web server from file syncing. Rewrote Dockerfile
