--- datoxidize/example_dirp1.csv
@@ -1,4 +1,0 @@
-ctenanthe setosa,-22.956208,-43.032058
-ctenanthe setosa,-0.1994138889,-78.4320916667
-lophostemon confertus,37.48782,-122.2383033333
-lophostemon confertus,-25.8574265897,152.6271105185
