+++ server/.env
@@ -1,0 +1,9 @@
+
+# root persistent storage for webserver
+ROOT_STORAGE=./storage/
+
+# testing directory
+#TEST_DIRECTORY=./example_dir/
+
+LOCAL_IP=[127,0,0,0]
+DOCKER_IP=[0,0,0,0]
