+++ server/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rayon="1.6.0"
+serde = { version = "1.0.148", default-features = false, features = ["derive"] }
+serde_json = "1.0.89"
+dotenvy = "0.15.6"
+tokio = { version = "1.22.0", features = ["full"] }
+rustls = "0.20.7"
+notify = "5.0.0"
+axum = "0.6.1"
+tracing = "0.1.37"
+tracing-subscriber = "0.3.16"
+rand = "0.8.4"
+askama = "0.11.0"
+fs_extra = "1.2.0"
+serial_test = "0.10.0"
